+++ openapydantic/versions/openapi_302/models.py
@@ -1,233 +1,43 @@
-import copy
 import enum
 import typing as t
 
 import pydantic
-from jsonpath_ng import parse
 
-import openapydantic
+from openapydantic import common
+from openapydantic import resolver
 
-ComponentType = openapydantic.common.ComponentType
-HTTPStatusCode = openapydantic.common.HTTPStatusCode
-MediaType = openapydantic.common.MediaType
-OpenApiBaseModel = openapydantic.common.OpenApiBaseModel
+HTTPStatusCode = common.HTTPStatusCode
+MediaType = common.MediaType
+OpenApiBaseModel = common.OpenApiBaseModel
+ComponentType = common.ComponentType
 
 Field = pydantic.Field
 
 
-def _get_ref_data(
-    ref: str,
-) -> t.Tuple[ComponentType, str]:
-    ref_split = ref.split("/")  # format #/components/schemas/Pet
-    ref_type = ComponentType(ref_split[2])
-    ref_key = ref_split[-1]
-    return ref_type, ref_key
-
-
-class ComponentsResolver:
-    with_ref: t.Dict[str, t.Any] = {}
-    without_ref: t.Dict[str, t.Any] = {}
-    ref_find = False
-    consolidate_count = 0
-    self_ref: t.List[str] = []
-
-    @classmethod
-    def init(cls):
-        cls.with_ref = {}
-        cls.without_ref = {}
-        cls.ref_find = False
-        cls.consolidate_count = 0
-
-        for elt in ComponentType:
-            cls.with_ref[elt.name] = {}
-            cls.without_ref[elt.name] = {}
-            cls.self_ref = []
-
-    @staticmethod
-    def _validate_references_format(
-        references: t.List[str],
-    ) -> None:
-        for ref in references:
-            if ".yaml" in ref or ".json" in ref:
-                raise NotImplementedError("File reference not implemented")
-            if not ref.startswith("#/"):
-                raise ValueError(f"reference {ref} has invalid format")
-
-    @classmethod
-    def _list_self_references(
-        cls,
-        *,
-        key: str,
-        component_type: ComponentType,
-        references: t.List[str],
-    ) -> None:
-        for ref in references:
-            ref_type, ref_key = _get_ref_data(ref=ref)
-            if ref_type == component_type and ref_key == key:
-                cls.self_ref.append(ref)
-
-    @classmethod
-    def _find_ref(
-        cls,
-        *,
-        obj: t.Any,
-    ) -> t.List[str]:
-        jsonpath_expr = parse("$..'$ref'")  # type:ignore
-        return list(
-            set([match.value for match in jsonpath_expr.find(obj)])  # type:ignore
+def reference_interpolation(
+    values: t.Dict[str, t.Any],
+) -> t.Dict[str, t.Any]:
+    ref = values.get("$ref")
+    # print("====== VALIDATE ROOT ======")
+    # print(f"ref:{ref}")
+    if ref:
+        # Avoir self reference here
+        if ref in resolver.ComponentsResolver.self_ref:
+            return values
+        ref_type, ref_key = resolver.get_ref_data(
+            ref=ref,
         )
+        # print(f"ref_type:{ref_type.name}")
+        # print(f"ref_key:{ref_key}")
+        ref_found: t.Dict[str, t.Any] = resolver.ComponentsResolver.without_ref[
+            ref_type.name
+        ].get(ref_key)
+        # print(f"ref_found:{ref_found}")
+        if not ref_found:
+            raise ValueError(f"Reference not found:{ref_type}/{ref_key}")
 
-    @classmethod
-    def _search_component_for_ref(
-        cls,
-        *,
-        key: str,
-        value: t.Dict[str, t.Any],
-        component_type: ComponentType,
-    ):
-        references = cls._find_ref(
-            obj=value,
-        )
-
-        cls.ref_find = bool(references)
-
-        cls._validate_references_format(
-            references=references,
-        )
-
-        cls._list_self_references(
-            key=key,
-            component_type=component_type,
-            references=references,
-        )
-
-        if references:
-            cls.with_ref[component_type.name][key] = {}
-            cls.with_ref[component_type.name][key]["values"] = value
-            cls.with_ref[component_type.name][key]["references"] = references
-        else:
-            cls.without_ref[component_type.name][key] = value
-
-    @classmethod
-    def _search_components_for_ref(
-        cls,
-        *,
-        components: t.Dict[str, t.Any],
-        component_type: ComponentType,
-    ):
-        for key, value in components.items():
-            cls._search_component_for_ref(
-                key=key,
-                value=value,
-                component_type=component_type,
-            )
-
-    @staticmethod
-    def _get_component_object(
-        component_type: ComponentType,
-        values: t.Dict[str, t.Any],
-    ) -> t.Dict[str, t.Any]:
-        if component_type == ComponentType.schemas:
-            component = Schema(**values)
-        elif component_type == ComponentType.headers:
-            component = Header(**values)
-        elif component_type == ComponentType.responses:
-            component = Response(**values)
-        elif component_type == ComponentType.parameters:
-            component = Parameter(**values)
-        elif component_type == ComponentType.examples:
-            component = Example(**values)
-        elif component_type == ComponentType.request_bodies:
-            component = RequestBody(**values)
-        elif component_type == ComponentType.links:
-            component = Link(**values)
-        elif component_type == ComponentType.callbacks:
-            component = PathItem(**values)
-
-        return component.as_clean_dict()
-
-    @classmethod
-    def _references_availables(
-        cls,
-        references: t.List[str],
-        component_type: ComponentType,
-        key: str,
-    ) -> bool:
-
-        for ref in references:
-            ref_type, ref_key = _get_ref_data(ref=ref)
-
-            if ref_type == component_type and key == ref_key:
-                continue
-
-            if not cls.without_ref[ref_type.name].get(ref_key):
-                # print(f"ref unavailable: {ref}")
-                return False
-            # print(f"ref available: {ref}")
-        return True
-
-    @classmethod
-    def _consolidate_components(
-        cls,
-        component_type: ComponentType,
-    ) -> None:
-        cls.consolidate_count = cls.consolidate_count + 1
-        component_dict = {}
-        with_ref_copy = copy.deepcopy(cls.with_ref[component_type.name])
-
-        for key, values in with_ref_copy.items():
-            references = values["references"]
-
-            # print(f"Trying to create {component_type.name}:{key}")
-            if not cls._references_availables(
-                references=references,
-                key=key,
-                component_type=component_type,
-            ):
-                # print(f"Not all references ready for:{component_type.name}:{key}")
-                # print("next...")
-                continue
-
-            component_dict = cls._get_component_object(
-                component_type=component_type,
-                values=values["values"],
-            )
-
-            cls.without_ref[component_type.name][key] = component_dict
-            del cls.with_ref[component_type.name][key]
-
-        if len(cls.with_ref[component_type.name]):
-            cls._consolidate_components(component_type=component_type)
-
-    @classmethod
-    def resolve(
-        cls,
-        *,
-        raw_api: t.Dict[str, t.Any],
-    ) -> None:
-        cls.init()
-
-        components = raw_api.get("components")
-        if not components:
-            # print("No components in this api")
-            return
-
-        for elt in ComponentType:
-            component = components.get(elt.value)
-            if component:
-                cls._search_components_for_ref(
-                    components=component,
-                    component_type=elt,
-                )
-
-        for elt in ComponentType:
-            component = components.get(elt.value)
-            if component:
-                cls._consolidate_components(component_type=elt)
-            # print(
-            #     f"Recursive consolidate count for component {elt.name}:"
-            #     f"{cls.consolidate_count}"
-            # )
+        return ref_found
+    return values
 
 
 class RefModel(OpenApiBaseModel):
@@ -240,30 +50,13 @@ class RefModel(OpenApiBaseModel):
         pre=True,
         allow_reuse=True,
     )
-    def validate_root(
+    def _reference_interpolation(
         cls,
         values: t.Dict[str, str],
     ) -> t.Dict[str, t.Any]:
-        ref = values.get("$ref")
-        # print("====== VALIDATE ROOT ======")
-        # print(f"ref:{ref}")
-        if ref:
-            # Avoir self reference here
-            if ref in ComponentsResolver.self_ref:
-                return values
-            # load reference from ComponentsResolver
-            ref_type, ref_key = _get_ref_data(ref)
-            # print(f"ref_type:{ref_type.name}")
-            # print(f"ref_key:{ref_key}")
-            ref_found: t.Dict[str, t.Any] = ComponentsResolver.without_ref[
-                ref_type.name
-            ].get(ref_key)
-            # print(f"ref_found:{ref_found}")
-            if not ref_found:
-                raise ValueError(f"Reference not found:{ref_type}/{ref_key}")
-
-            return ref_found
-        return values
+        return reference_interpolation(
+            values=values,
+        )
 
 
 class BaseModelForbid(RefModel):